Use double for MenuModel.Price and add tests
Change MenuModel.Price from decimal to double (including constructor) and update ManageMenu to use double variables and double.TryParse for price input and validation. Add ReservationSysteemTesting/ManageMenuTest.cs with basic unit tests covering MenuLogic operations (create menu, add/update menu items, and delete operations).

+++ ReservationSysteem/DataModels/MenuModel.cs
@@ -4,7 +4,7 @@ public class MenuModel
     public string MenuName { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public decimal Price { get; set; }
+    public double Price { get; set; }
     public string FoodCategory { get; set; }
     public string Allergens { get; set; }
     public bool IsActive { get; set; }
@@ -13,7 +13,7 @@ public class MenuModel
     {
     }
 
-    public MenuModel(string menuName, string name, string description, decimal price, string foodCategory, string allergens)
+    public MenuModel(string menuName, string name, string description, double price, string foodCategory, string allergens)
     {
         MenuName = menuName;
         Name = name;
